Clarify rspec running
Remove the suggestion to directly run rspec as it doesn't work on Java 17 due to special args that are needed for the server. Alternative would be to add the required GoCD Server "opens" to the jruby script template, but then they would be there for all JRuby usages which is probably not a good idea; might cause things to work when they would not in a "real" environment. Needs some more thought.